[gotemp] Use '%.2f' format instead of '%g'

+++ src/PCsensor_Temper.c
@@ -19,6 +19,7 @@
 
 #include <stdio.h>
 #include <stdlib.h>
+#include <stdint.h>
 #include <string.h>
 #include <errno.h>
 
@@ -106,9 +107,9 @@ static int PCsensor_Temper_update(struct usense_device *dev, void *priv)
 	} else {
 		/* Kelvin */
 		char buff[48];
-		double celsius = (temp / 256.0 + CAL_ADD) * CAL_MUL;
+		double celsius = ((double)temp / 256.0 + CAL_ADD) * CAL_MUL;
 		double kelvin = C_TO_K(celsius);
-		snprintf(buff, sizeof(buff), "%.2g", kelvin);
+		snprintf(buff, sizeof(buff), "%.2f", kelvin);
 		usense_prop_set(dev, "reading", buff);
 	}
